+++ examples/react/AnimatedHeader/src/data/index.jsx
@@ -8,27 +8,29 @@
  */
 
 import { Add } from '@carbon/react/icons';
-import { ButtonKinds } from '@carbon/react';
+import { ButtonKinds, Loading } from '@carbon/react';
+import SampleCustomTaskContent from './SampleCustomTaskContent';
+import React from 'react';
 
 export const workspaceData = [
   {
     id: 'workspace-1',
-    text: 'Workspace 1',
+    label: 'Workspace 1',
   },
   {
     id: 'workspace-2',
-    text: 'Workspace 2',
+    label: 'Workspace 2',
   },
   {
     id: 'workspace-3',
-    text: 'Workspace 3',
+    label: 'Workspace 3',
   },
 ];
 
 export const headerTiles = [
   {
     id: 1,
-    name: 'AI Chat Tile w/ two glass tiles',
+    label: 'AI Chat Tile w/ two glass tiles',
     tiles: [
       {
         id: 'ai-tile',
@@ -54,7 +56,7 @@ export const headerTiles = [
   },
   {
     id: 2,
-    name: 'Three glass tiles',
+    label: 'Three glass tiles',
     tiles: [
       {
         id: 'tile-1',
@@ -81,7 +83,7 @@ export const headerTiles = [
   },
   {
     id: 3,
-    name: 'Four glass tiles',
+    label: 'Four glass tiles',
     tiles: [
       {
         id: 'tile-1',
@@ -115,7 +117,7 @@ export const headerTiles = [
   },
   {
     id: 4,
-    name: 'Five glass tiles',
+    label: 'Five glass tiles',
     tiles: [
       {
         id: 'tile-1',
@@ -156,7 +158,7 @@ export const headerTiles = [
   },
   {
     id: 5,
-    name: 'Six glass tiles',
+    label: 'Six glass tiles',
     tiles: [
       {
         id: 'tile-1',
@@ -204,7 +206,7 @@ export const headerTiles = [
   },
   {
     id: 6,
-    name: 'Seven glass tiles',
+    label: 'Seven glass tiles',
     tiles: [
       {
         id: 'tile-1',
@@ -259,7 +261,7 @@ export const headerTiles = [
   },
   {
     id: 7,
-    name: 'Eight glass tiles',
+    label: 'Eight glass tiles',
     tiles: [
       {
         id: 'tile-1',
@@ -316,6 +318,20 @@ export const headerTiles = [
         title: 'Create and run python queries',
         subtitle: 'with Prompt Lab',
         mainIcon: 'DataSet',
+      },
+    ],
+  },
+  {
+    id: 8,
+    label: 'Custom content tasks',
+    tiles: [
+      {
+        id: 'tile-1',
+        customContent: <Loading description="Sample loading state" />
+      },
+      {
+        id: 'tile-2',
+        customContent: <SampleCustomTaskContent />
       },
     ],
   },