+++ packages/react/src/components/AnimatedHeader/CHANGELOG.md
@@ -3,6 +3,77 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.10.0...@carbon-labs/react-animated-header@0.11.0) (2025-04-18)
+
+**Note:** Version bump only for package @carbon-labs/react-animated-header
+
+
+
+
+
+# [0.10.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.9.0...@carbon-labs/react-animated-header@0.10.0) (2025-04-10)
+
+
+### Bug Fixes
+
+* **animated-header:** add backup for autoprefixer ([#563](https://github.com/carbon-design-system/carbon-labs/issues/563)) ([8c3c4a6](https://github.com/carbon-design-system/carbon-labs/commit/8c3c4a6846627b6fbb8e65e50bb731acb52784dc))
+
+
+
+
+
+# [0.9.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.8.0...@carbon-labs/react-animated-header@0.9.0) (2025-04-10)
+
+
+### Bug Fixes
+
+* **animated-header:** cleanup Safari issues ([#556](https://github.com/carbon-design-system/carbon-labs/issues/556)) ([530e01a](https://github.com/carbon-design-system/carbon-labs/commit/530e01ae13d94391732bf34a3b26edb87b9ea161))
+
+
+
+
+
+# [0.8.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.7.0...@carbon-labs/react-animated-header@0.8.0) (2025-04-08)
+
+
+### Features
+
+* **animated-header:** remove strictmode from example app + tooltip ([#548](https://github.com/carbon-design-system/carbon-labs/issues/548)) ([c969903](https://github.com/carbon-design-system/carbon-labs/commit/c969903eab1bd7b78fbb3b7f9770e4839b1d7f6b))
+
+
+
+
+
+# [0.7.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.6.0...@carbon-labs/react-animated-header@0.7.0) (2025-04-04)
+
+**Note:** Version bump only for package @carbon-labs/react-animated-header
+
+
+
+
+
+# [0.6.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.5.0...@carbon-labs/react-animated-header@0.6.0) (2025-04-01)
+
+
+### Features
+
+* **animated-header:** remove loop animation from header ([#547](https://github.com/carbon-design-system/carbon-labs/issues/547)) ([613ffc3](https://github.com/carbon-design-system/carbon-labs/commit/613ffc358e22b8270dc0535cd6d5d4a352dc8ac8))
+
+
+
+
+
+# [0.5.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.4.0...@carbon-labs/react-animated-header@0.5.0) (2025-04-01)
+
+
+### Features
+
+* **animated-header:** add custom itemToString func for Dropdowns ([#542](https://github.com/carbon-design-system/carbon-labs/issues/542)) ([8ae02cf](https://github.com/carbon-design-system/carbon-labs/commit/8ae02cf012adec82e3fd962bf3f84df0752023e9))
+
+
+
+
+
 # [0.4.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-animated-header@0.3.0...@carbon-labs/react-animated-header@0.4.0) (2025-03-26)